Merge 1.4 to master

+++ src/ua_types_encoding_json.h
@@ -59,8 +59,8 @@ UA_StatusCode writeJsonBeforeElement(CtxJson *ctx, UA_Boolean distinct);
 typedef struct {
     const char *json5;
     cj5_token *tokens;
-    unsigned int tokensSize;
-    unsigned int index;
+    size_t tokensSize;
+    size_t index;
     UA_Byte depth;
 
     size_t namespacesSize;
